Switch to a package structure
