+++ dotrix_core/src/animation.rs
