+++ dotrix_core/src/application.rs
