+++ dotrix_core/src/assets.rs
