+++ dotrix_core/src/assets/animation.rs
