+++ dotrix_core/src/assets/id.rs
