+++ dotrix_core/src/assets/load_gltf.rs
