+++ dotrix_core/src/assets/loader.rs
