+++ dotrix_core/src/assets/mesh.rs
