+++ dotrix_core/src/assets/resource.rs
