+++ dotrix_core/src/assets/skin.rs
