+++ dotrix_core/src/camera.rs
