+++ dotrix_core/src/renderer.rs
