+++ dotrix_core/src/renderer/light.rs
