+++ dotrix_core/src/renderer/model.rs
