+++ dotrix_core/src/renderer/pipeline.rs
