+++ dotrix_core/src/renderer/shaders/skinned.frag.spv
