+++ dotrix_core/src/renderer/shaders/skinned.vert.spv
