+++ dotrix_core/src/renderer/shaders/skybox.frag.spv
