+++ dotrix_core/src/renderer/shaders/skybox.vert.spv
