+++ dotrix_core/src/renderer/shaders/static.frag.glsl
