+++ dotrix_core/src/renderer/shaders/static.vert.glsl
