+++ dotrix_core/src/renderer/skybox.rs
