+++ dotrix_ui/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "dotrix_ui"
+version = "0.2.0"
+authors = [
+  "Elias Kartashov <elias@lowenware.com>",
+  "Štěpán Wünsch <sw@lowenware.com>",
+]
+edition = "2018"
+description = "UI for Dotrix 3D game engine"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dotrix_core = { version = "0.2", path = "../dotrix_core" }
