+++ dotrix_ui/src/lib.rs
@@ -1,0 +1,5 @@
+use dotrix_core::math::Transform;
+
+pub fn some_ui_provider() -> Transform {
+    Transform::from_scale_factor(2.0)
+}
