+++ examples/ui/ui.rs
@@ -1,0 +1,5 @@
+use dotrix::ui::some_ui_provider;
+
+fn main() {
+    println!("{:?}", some_ui_provider());
+}
